+++ src/utils/price.js
@@ -1,23 +1,12 @@
-import {
-  MENUS,
-  BENEFIT_PRICE,
-  DECEMBER,
-  WEEKENDS,
-  SPECIAL_DAYS,
-  DESSERTS,
-  MAIN_MENUS,
-  BADGES,
-  NOTHING_MESSAGE,
-} from "./constants/index.js";
+import { MENUS, MAIN_MENUS, DESSERTS, BENEFIT_PRICE, DECEMBER } from "../constants/index.js";
+import { getCategoryMenuCount } from "./event.js";
 
+// 할인 전 가격 금액
 export const getOriginalOrderPrice = (orderMenus) => {
   return orderMenus.reduce((acc, { menu, count }) => acc + MENUS[menu] * count, 0);
 };
 
-export const isGiftEventTarget = (originalPrice) => {
-  return originalPrice >= BENEFIT_PRICE.GIFT;
-};
-
+// 크리스마스 디데이 할인 금액
 export const getChristmasEventDiscount = (visitDate) => {
   if (visitDate > DECEMBER.CHRISTMAS) {
     return 0;
@@ -26,24 +15,7 @@ export const getChristmasEventDiscount = (visitDate) => {
   return (visitDate - 1) * BENEFIT_PRICE.HUNDRED + BENEFIT_PRICE.THOUSAND;
 };
 
-export const isWeekend = (visitDate) => {
-  return WEEKENDS.includes(visitDate);
-};
-
-export const getCategoryMenuCount = (categoryMenu, orderMenu) => {
-  return orderMenu.reduce((acc, { menu, count }) => {
-    if (categoryMenu.includes(menu)) {
-      return acc + count;
-    }
-
-    return acc;
-  }, 0);
-};
-
-export const isSpecialEventDay = (visitDate) => {
-  return SPECIAL_DAYS.includes(visitDate);
-};
-
+// 평일, 주말 할인 금액
 export const getDayEventDiscount = (isWeekend, orderMenus) => {
   if (isWeekend) {
     const mainMenuCount = getCategoryMenuCount(MAIN_MENUS, orderMenus);
@@ -54,6 +26,7 @@ export const getDayEventDiscount = (isWeekend, orderMenus) => {
   return dessertMenuCount * BENEFIT_PRICE.YEAR;
 };
 
+// 총 할인 금액
 export const getDiscountPrice = (christmasDiscount, dayDiscount, isGiftEvent, isSpecialDate) => {
   let discountPrice = 0;
   discountPrice += christmasDiscount;
@@ -70,22 +43,7 @@ export const getDiscountPrice = (christmasDiscount, dayDiscount, isGiftEvent, is
   return discountPrice;
 };
 
+// 최종 결제 금액
 export const getTotalPrice = (originalPrice, discountPrice) => {
   return originalPrice - discountPrice;
-};
-
-export const getBadge = (discountPrice) => {
-  if (discountPrice >= BENEFIT_PRICE.TWENTY_THOUSAND) {
-    return BADGES.SANTA;
-  }
-
-  if (discountPrice >= BENEFIT_PRICE.TEN_THOUSAND) {
-    return BADGES.TREE;
-  }
-
-  if (discountPrice >= BENEFIT_PRICE.FIVE_THOUSAND) {
-    return BADGES.STAR;
-  }
-
-  return NOTHING_MESSAGE;
 };
